Modified WBP_CharacterUI
Updated the WBP_CharacterUI Widget so that it now includes the T_Full_Revenant Texture as a visual representation of the Revenant character.  Added descriptive text to display the character's name and modified the font value.  Modified the Image_Background so that it now shows the T_MinionSmokeTile Texture.  Set all corresponding anchor points and relative position values.