Player Status Layout
Modified the WBP_CharacterUI Widget so that it now includes a layout for Player Status, which includes the player's level, money, and XP.  Implemented using nested HBox and ProgressBar controls within a parent VBox.  Changed the Image_Money widget to display the T_Icon_Money Texture as an Image brush.