Character Roster DT & Export
Created the DT_CharacterRoster Data Table and set the Row Structure to be STR_CharacterData.  Modified the Data Table to include rows for Revenant, Muriel, and Sevarog.  Used the DT_CharacterRoster asset to export out into the DT_CharacterRoster CSV file.  Added the CSV file to the CharacterUI/Data subdirectory.