Default Selected Ability Script
Modified WBP_CharacterUI so that the Button_Ability controls now feature a yellow tint on hover.  Also modified the Widget's Event Graph to include the custom event "Default Ability Selected," which provides Button_Ability and Animation instance references for the Hovered Button FX function and sets the value of SelectedAbilityIndex.  Added a Sequence node after Event Construct to set values for BPGameInstanceRef as well as the default SelectedAbilityIndex.