Character Name Attribute Input
Modified WBP_CharacterUI so that the CanvasPanel_Attributes Scroll Box now features a nested Horizontal Box for inputting the character's Name attribute.  Added the attribute descriptor Text_CharacterName, anchored it to the left side of the HBox, and set values for font family, outline size, shadow offset, and shadow color.  Added an Editable Text Box, anchored it to the right of the HBox, and set values for size, horizontal alignment, font family, font typeface, and font size.