Health Attribute Progress Bar
Modified WBP_CharacterUI so that the CanvasPanel_Attribute Scroll Box now features an attribute descriptor for the character's Health.  Added a Progress Bar, aligned it to the right of the HBox, and set values for padding, tint, and fill color.  Added an additional Text control for displaying the Health's numeric total and set values for text, size, and horizontal alignment.