Abilities / Attributes Selection Script
Modified WBP_CharacterUI so that the event graph now features OnClicked selection scripts for the Abilities and Attributes Buttons, which sets the active widget index value for the WidgetSwitcher_Categories control, sets the render scale and background color values for the clicked button, and sets default render scale and background color values for the button that was not clicked.  Also added a new custom event, "Default Category Selected," which sets the default active index value for the WidgetSwitcher_Categories control.