Abilities / Attributes Gamepad Script
Modified WBP_CharacterUI so that the event graph now features a new custom event, "Abilities Tab Selected," which executes the same node network as the OnClicked event for Button_Category_01.  Added a second custom event, "Attributes Tab Selected," which executes the same node network as the OnClicked event for Button_Category_02.  Modified BP_PlayerController's Event Graph so that it now includes a node network for calling AbilitiesTabSelected on the GamepadLeftShoulder event, and AttributesTabSelected on the GamepadRightShoulder event.  Set the Execute When Paused value to true for both the GamepadLeftShoulder and GamepadRightShoulder events.