Auto Move Speed Data Binding
Modified WBP_CharacterUI so that the HBox_AutoMoveSpeed control now features a binding function for the Is Enabled property.  Modified Slider_AutoMoveSpeed to include a binding function for the Value property.  Modified Text_AutoMoveSpeed to include a binding function for the Text property.  Modified DT_CharacterRoster_Full to include updated Auto Move Speed values for Revenant, Muriel, and Sevarog.