Name Attribute Save Script
Modified BP_GameInstance so that it now features a new function, "UpdateAttributeName," which gets a reference to the CharacterData_Inst element at the current PlayerInstance index and uses a Set Array Elem node to copy all values from the current element, plus the updated value for the Name attribute.  Also modified WBP_CharacterUI so that it now features implemented node networks for EditableTextBox_Name's OnTextChanged and OnTextCommitted events.